+++ upload/catalog/controller/extension/payment/emspay_ideal.php
@@ -129,7 +129,7 @@ class ControllerExtensionPaymentEmspayIdeal extends Controller
      */
     protected function createOrder(array $orderData)
     {
-        return $this->ems->createOrder([
+        return $this->ems->createOrder(array_filter([
             'amount' => $orderData['amount'],                                // Amount in cents
             'currency' => $orderData['currency'],                            // Currency
             'description' => $orderData['description'],                      // Description
@@ -144,6 +144,6 @@ class ControllerExtensionPaymentEmspayIdeal extends Controller
                     'payment_method_details' => ['issuer_id' => $orderData['issuer_id']]
                 ]
             ]
-        ]);
+        ]));
     }
 }