+++ upload/catalog/controller/extension/payment/emspay_klarnapaylater.php
@@ -173,7 +173,7 @@ class ControllerExtensionPaymentEmspayKlarnaPayLater extends Controller
      */
     protected function createOrder(array $orderData)
     {
-        return $this->ems->createOrder([
+        return $this->ems->createOrder(array_filter([
             'amount' => $orderData['amount'],                                // Amount in cents
             'currency' => $orderData['currency'],                            // Currency
             'description' => $orderData['description'],                      // Description
@@ -188,6 +188,6 @@ class ControllerExtensionPaymentEmspayKlarnaPayLater extends Controller
                     'payment_method' => "klarna-pay-later",
                 ]
             ]
-        ]);
+        ]));
     }
 }